Rename 패키지 명 변경 web -> app
