+++ src/main/java/yapp/buddycon/app/gifticon/adapter/GifticonController.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.web.gifticon.adapter;
+package yapp.buddycon.app.gifticon.adapter;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -6,9 +6,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import yapp.buddycon.web.gifticon.adapter.request.SearchGifticonDTO;
-import yapp.buddycon.web.gifticon.adapter.response.GifticonResponseDTO;
-import yapp.buddycon.web.gifticon.application.port.in.GifticonUseCase;
+import yapp.buddycon.app.gifticon.adapter.request.SearchGifticonDTO;
+import yapp.buddycon.app.gifticon.adapter.response.GifticonResponseDTO;
+import yapp.buddycon.app.gifticon.application.port.in.GifticonUseCase;
 
 @RestController
 @RequiredArgsConstructor
