+++ src/main/java/yapp/buddycon/app/gifticon/application/service/GifticonService.java
@@ -1,14 +1,14 @@
-package yapp.buddycon.web.gifticon.application.service;
+package yapp.buddycon.app.gifticon.application.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import yapp.buddycon.web.gifticon.adapter.request.SearchGifticonDTO;
-import yapp.buddycon.web.gifticon.adapter.response.GifticonResponseDTO;
-import yapp.buddycon.web.gifticon.application.port.in.GifticonUseCase;
-import yapp.buddycon.web.gifticon.application.port.out.GifticonQueryPort;
+import yapp.buddycon.app.gifticon.adapter.request.SearchGifticonDTO;
+import yapp.buddycon.app.gifticon.adapter.response.GifticonResponseDTO;
+import yapp.buddycon.app.gifticon.application.port.in.GifticonUseCase;
+import yapp.buddycon.app.gifticon.application.port.out.GifticonQueryPort;
 
 @Service
 @RequiredArgsConstructor
