+++ src/main/java/yapp/buddycon/app/gifticon/domain/StoreCategory.java
@@ -1,12 +1,10 @@
-package yapp.buddycon.web.gifticon.domain;
+package yapp.buddycon.app.gifticon.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -18,17 +16,13 @@ import yapp.buddycon.common.BaseEntity;
 @AllArgsConstructor
 @Builder
 @Entity
-public class Store extends BaseEntity {
+public class StoreCategory extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "store_id")
+  @Column(name = "store_category_id")
   private Long id;
   @Column(name = "name", nullable = false, length = 20)
   private String name;
-
-  @ManyToOne
-  @JoinColumn(name = "store_category_id", nullable = false, referencedColumnName = "store_category_id")
-  private StoreCategory storeCategory;
 
 }
