+++ src/main/java/yapp/buddycon/app/gifticon/infra/GifticonRepository.java
@@ -1,13 +1,13 @@
-package yapp.buddycon.web.gifticon.infra;
+package yapp.buddycon.app.gifticon.infra;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
-import yapp.buddycon.web.gifticon.adapter.request.SearchGifticonDTO;
-import yapp.buddycon.web.gifticon.adapter.response.GifticonResponseDTO;
-import yapp.buddycon.web.gifticon.infra.jpa.GifticonJpaRepository;
-import yapp.buddycon.web.gifticon.application.port.out.GifticonQueryPort;
+import yapp.buddycon.app.gifticon.adapter.request.SearchGifticonDTO;
+import yapp.buddycon.app.gifticon.adapter.response.GifticonResponseDTO;
+import yapp.buddycon.app.gifticon.infra.jpa.GifticonJpaRepository;
+import yapp.buddycon.app.gifticon.application.port.out.GifticonQueryPort;
 
 @RequiredArgsConstructor
 @Repository
