--- src/main/java/yapp/buddycon/web/gifticon/application/port/in/GifticonUseCase.java
@@ -1,11 +1,0 @@
-package yapp.buddycon.web.gifticon.application.port.in;
-
-import org.springframework.data.domain.Page;
-import yapp.buddycon.web.gifticon.adapter.request.SearchGifticonDTO;
-import yapp.buddycon.web.gifticon.adapter.response.GifticonResponseDTO;
-
-public interface GifticonUseCase {
-
-  Page<GifticonResponseDTO> getGifticons(SearchGifticonDTO dto);
-
-}
